requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
spacy==3.7.5
python-dotenv==1.0.1
google-generativeai==0.8.3
streamlit==1.39.0
requests==2.32.3
